+++ src/page-objects/administration/CustomFieldDetail.ts
@@ -1,7 +1,7 @@
-import type { Page, Locator } from '@playwright/test';
+import type { Page, Locator } from 'playwright-core';
 import { CustomFieldCreate } from './CustomFieldCreate';
 import { getSelectFieldListitem } from './modules/SelectFieldListitem';
-import { HelperFixtureTypes } from '../../fixtures/HelperFixtures';
+import type { HelperFixtureTypes } from '../../fixtures/HelperFixtures';
 import { satisfies } from 'compare-versions';
 
 export class CustomFieldDetail extends CustomFieldCreate {
@@ -23,9 +23,11 @@ export class CustomFieldDetail extends CustomFieldCreate {
     public readonly customFieldDeleteCancelButton: Locator;
     public readonly customFieldDeleteButton: Locator;
     public readonly customFieldEditAvailableInShoppingCartCheckbox: Locator;
+    public readonly instanceMeta: HelperFixtureTypes['InstanceMeta'];
 
-    constructor(public readonly page: Page, public readonly instanceMeta: HelperFixtureTypes['InstanceMeta']) {
+    constructor(page: Page, instanceMeta: HelperFixtureTypes['InstanceMeta']) {
         super(page);
+        this.instanceMeta = instanceMeta;
 
         //Custom field section
         this.newCustomFieldButton = page.getByRole('button', { name: 'New custom field' });