+++ src/tasks/shop-admin/Rule/AssignEntitiesToRule.ts
@@ -1,8 +1,9 @@
-import { Locator, test as base } from '@playwright/test';
+import { test as base } from '@playwright/test';
+import type { Locator } from 'playwright-core';
 import type { Task } from '../../../types/Task';
 import type { FixtureTypes } from '../../../types/FixtureTypes';
 import { RuleType } from '../../../types/ShopwareTypes';
-import { RuleAssignmentEntity } from '../../../types/ShopwareTypes';
+import type { RuleAssignmentEntity } from '../../../types/ShopwareTypes';
 
 export const AssignEntitiesToRule = base.extend<{ AssignEntitiesToRule: Task }, FixtureTypes>({
     AssignEntitiesToRule: async ({ AdminRuleDetail }, use ) => {