+++ src/tasks/shop-customer/Product/SelectProductFilterOption.ts
@@ -1,4 +1,5 @@
-import { test as base, Locator } from '@playwright/test';
+import { test as base } from '@playwright/test';
+import type { Locator } from 'playwright-core';
 import type { Task } from '../../../types/Task';
 import type { FixtureTypes } from '../../../types/FixtureTypes';